Install dependencies automatically when installing this package
Previously, it would crash if markdown is not installed, because mdx_latex imports markdown
Solved by moving dhe __version__ and __doc__ to a separate file that does not import markdown
